move typeinfo_string function to functions
